+++ instance.tf
@@ -40,7 +40,4 @@ resource "proxmox_vm_qemu" "proxmox_instance" {
   cicustom = var.pxe_boot == true ? null : var.cicustom
   # example: "user=${var.citemplate_storage}:${var.snippet_dir}/user-${var.snippet_file_base},network=${var.citemplate_storage}:${var.snippet_dir}/network-${var.snippet_file_base}"
   cloudinit_cdrom_storage = var.pxe_boot == true ? null : var.cloudinit_cdrom_storage
-
-  searchdomain = var.searchdomain
-  nameserver   = var.nameserver
 }