+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Remove `searchdomain` and `nameserver` vars.
+
 ## [1.2.2] - 2022-11-27
 
 ## [1.2.1] - 2022-11-27